+++ roadeye-api/roadeye-api-mdt-log-collector/src/main/java/org/re/mdtlog/collector/app/common/dto/BaseMdtLogResponse.java
@@ -1,4 +1,4 @@
-package org.re.mdtlog.collector.app.common;
+package org.re.mdtlog.collector.app.common.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import org.re.mdtlog.collector.exception.MdtLogExceptionCode;
