+++ roadeye-api/roadeye-api-mdt-log-collector/src/main/java/org/re/mdtlog/collector/app/cyclelog/MdtCycleLogController.java
@@ -2,7 +2,7 @@ package org.re.mdtlog.collector.app.cyclelog;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.re.mdtlog.collector.app.common.BaseMdtLogResponse;
+import org.re.mdtlog.collector.app.common.dto.BaseMdtLogResponse;
 import org.re.mdtlog.collector.app.cyclelog.dto.MdtAddCycleLogRequest;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
